rename: ToolTip web domain common으로 이동
