+++ packages/web-domains/src/common/components/ToolTip/ToolTip.styles.ts
